% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/D9_data.R
\docType{data}
\name{D9_data}
\alias{D9_data}
\title{D9 Artificial Dataset}
\format{
A data frame with 1000 data points and 3 variables
\describe{
  \item{\code{x}}{Numeric values generated from Uniform distributions}
  \item{\code{y}}{Numeric values generated from Uniform distributions}
  \item{\code{label}}{Categorical labels {1,2,3}}
}
}
\usage{
D9_data
}
\description{
A \code{2}-dimensional dataset from the FuzzyDatasets package (https://github.com/O-PREEDASAWAKUL/FuzzyDatasets) generated from \code{3} different Uniform distributions labeled as \code{1-3}.
}
\references{
{O. Preedasawakul and N. Wiroonsri, FuzzyDatasets, (2023), GitHub repository, \url{https://github.com/O-PREEDASAWAKUL/FuzzyDatasets}}
}
\seealso{
{\link{FzzyCVIs},  \link{WP.IDX},  \link{R1_data}}
}
\author{
{Nathakhun Wiroonsri and Onthada Preedasawakul}
}
\keyword{datasets}
