% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R1_data.R
\docType{data}
\name{R1_data}
\alias{R1_data}
\title{R1 Artificial Dataset}
\format{
A data frame with 450 data points and 3 variables
\describe{
  \item{\code{x}}{Numeric values generated from Gaussian distributions}
  \item{\code{y}}{Numeric values generated from Gaussian distributions}
  \item{\code{label}}{Categorical labels {1,2,3,4,5,6,7,8,9}}
}
}
\usage{
R1_data
}
\description{
A \code{2}-dimensional dataset from the FuzzyDatasets package (https://github.com/O-PREEDASAWAKUL/FuzzyDatasets) generated from \code{9} different Gaussian distributions labeled as \code{1-9}.
}
\references{
{O. Preedasawakul and N. Wiroonsri, FuzzyDatasets, (2023), GitHub repository, \url{https://github.com/O-PREEDASAWAKUL/FuzzyDatasets}}
}
\seealso{
{\link{FzzyCVIs},  \link{WP.IDX},  \link{D1_data}}
}
\author{
{Nathakhun Wiroonsri and Onthada Preedasawakul}
}
\keyword{datasets}
