% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R3_data.R
\docType{data}
\name{R3_data}
\alias{R3_data}
\title{R3 Artificial Dataset}
\format{
A data frame with 1600 data points and 3 variables
\describe{
  \item{\code{x}}{Numeric values generated from Gaussian distributions}
  \item{\code{y}}{Numeric values generated from Gaussian distributions}
  \item{\code{label}}{Categorical labels {1,2,3,...,16}}
}
}
\usage{
R3_data
}
\description{
A \code{2}-dimensional dataset from the FuzzyDatasets package (https://github.com/O-PREEDASAWAKUL/FuzzyDatasets) generated from \code{16} different Gaussian distributions labeled as \code{1-16}.
}
\references{
{O. Preedasawakul and N. Wiroonsri, FuzzyDatasets, (2023), GitHub repository, \url{https://github.com/O-PREEDASAWAKUL/FuzzyDatasets}}
}
\seealso{
{\link{FzzyCVIs},  \link{WP.IDX},  \link{D1_data}}
}
\author{
{Nathakhun Wiroonsri and Onthada Preedasawakul}
}
\keyword{datasets}
