% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R7_data.R
\docType{data}
\name{R7_data}
\alias{R7_data}
\title{R7 Artificial Dataset}
\format{
A data frame with 1200 data points and 3 variables
\describe{
  \item{\code{x}}{Numeric values generated from Gaussian and Uniform distributions}
  \item{\code{y}}{Numeric values generated from Gaussian and Uniform distributions}
  \item{\code{label}}{Categorical labels {1,2,3}}
}
}
\usage{
R7_data
}
\description{
A \code{2}-dimensional dataset from the FuzzyDatasets package (https://github.com/O-PREEDASAWAKUL/FuzzyDatasets) generated from \code{6} different Gaussian and \code{3} Uniform distributions labeled as \code{1-3}.
}
\references{
{O. Preedasawakul and N. Wiroonsri, FuzzyDatasets, (2023), GitHub repository, \url{https://github.com/O-PREEDASAWAKUL/FuzzyDatasets}}
}
\seealso{
{\link{FzzyCVIs},  \link{WP.IDX},  \link{D1_data}}
}
\author{
{Nathakhun Wiroonsri and Onthada Preedasawakul}
}
\keyword{datasets}
